+++ crates/ruff/resources/test/fixtures/pylint/repeated_equality_comparison.py
