+++ crates/ruff/src/rules/pylint/rules/mod.rs
@@ -33,7 +33,7 @@ pub(crate) use no_self_use::*;
 pub(crate) use nonlocal_without_binding::*;
 pub(crate) use property_with_parameters::*;
 pub(crate) use redefined_loop_name::*;
-pub(crate) use repeated_equality_comparison_target::*;
+pub(crate) use repeated_equality_comparison::*;
 pub(crate) use repeated_isinstance_calls::*;
 pub(crate) use return_in_init::*;
 pub(crate) use self_assigning_variable::*;
@@ -91,7 +91,7 @@ mod no_self_use;
 mod nonlocal_without_binding;
 mod property_with_parameters;
 mod redefined_loop_name;
-mod repeated_equality_comparison_target;
+mod repeated_equality_comparison;
 mod repeated_isinstance_calls;
 mod return_in_init;
 mod self_assigning_variable;
